dashboard/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
tweepy>=4.14.0
google-api-python-client>=2.100.0
schedule>=1.2.0